stage 2

+++ AsaryaBackEnd.Core/Models/Item.cs
@@ -2,10 +2,10 @@
 {
     public class Item
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
         public int? Barcode { get; set; }
         public string? Description { get; set; }
         public decimal? Price { get; set; }
         public int? Quantity { get; set; }
     }
-}
+}
